Remove queueing of undeliverable packets
The switch had a list of packets that could not be delivered. Instead of
dropping them, they were stored in a list to be delivered later, after a
timeout. Actually, a network stack will cope with dropped packets and
adapt itself to the available capacity. Thus remove the code to
handle pending packets.

Co-authored: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Change-Id: I91ce5c1b73655f0050dbeb9768cbb3dfaeac25cf
Requires-Change-Id: I9720b31827b051974773935bfb67acbec9520a08
Requires-Change-Id: Ib50b5dc0d1ead4633137fa1db5d1a600227fb3a6