Do not send packet to the port it came in
It may happen that a packet with a valid MAC address is sent via
broadcast, e.g. by a switch which does not know how to reach this
address and sends the packet to all ports, except the one it came in.

The virtual switch does the same when it gets a packet with an unknown
MAC address. But in the case of a known MAC address it sends the
packet to the port known to reach the target address even if it is
the same port the packet just came in. This can lead to two
virtio-net-switch instances generating large amount of network
packages.

Here we filter packets the would go out the port they came in (as we
do for broadcast packets which are not send to the port they came in).

Change-Id: Idff922f47ce1fa54887a3c259ae507968ff949fe