catch Bad_descriptor exceptions and reset device
Bad_descriptor exceptions are thrown, when the memory of the client
does not resemble a descriptor anymore. This can happen during a
memory transfer from source to destination.

An exception resets the source/destination device on which it ocurred.
Due to the fact that both ends throw the same exception type, we catch
destination exceptions directly and let source exception propagate
upwards to the handling routines, which must terminate when the source
goes bad.

Signaling a device_error resets the Port/Virtio_net device and
signals all other ports to remove pending requests from this source
port.

Since, we delete all pending requests on reset anyways, don't buffer
requests, when the port is not ready.

fixes ticket #CD-1303

Change-Id: I11c3425dc83dadf0fbb1080e67c717e086f75ab9