Add support for ixylon network ports
Integrates the virtio-net-switch with the Ixylon network driver library
to support hardware network ports.

To plug a hardware device into the switch, a Vbus capability with the
name `vbus` needs to be provided when starting the switch. The switch
will then detect the presence of a NIC supported by Ixylon on the Vbus
and automatically add it as a hardware network port to the switch.

This feature is gated by the `VNS_IXL` Kconfig option, which is disabled
by default.

Change-Id: If27994bd9dc7ad5bab8783fc9b70ac9a507e2bcc