l4virtio: Implement TX burst limit mechanism
The idea is to limit the number of consecutive TX requests a port can
process before being interrupted to ensure fairness to other ports.

When a port hits its TX burst limit, we stop processing TX work from
this port, and instead reschedule the pending work for later. This
rescheduling is implemented with a per-port software-triggered IRQ,
which gets triggered when hitting the TX burst limit, i.e. it gets
queued at the end of the list of outstanding messages to be processed
by the server loop.

Change-Id: I1dc98e77ffc0f5f4d4ca9c983b9983f7583e8bc2