[API] gracefully terminate in case of unknown cmdline arguments
To be in sync with virtio-p2p-link use 1 and change the exit code for
an invalid 'svr' capability to 2.

There is no need to throw an exception. Instead signal an unorderly
exit with a distinct exit value, which emits a signal via the
L4Re::Parent protocol. This signal is not sent, when the application
terminates with an exception.

Change-Id: I74223ce4d210d3870908d5896ecfd1697cbcd226