Implement port statistics monitoring
The factory of the switch can be used to create a capability to query
port statistics (type == 1). The protocol is defined in
<l4/virtio-net-switch/stats.h>. Ports are identified by the name
given by the "name=..." property when they were created.

The statistics interface provides access to the statistics of all ports
of the switch. The information is provided via shared memory, which is
accessible in a read-only manner to the client.

The counting of the port statistics can be selectively enabled
in Kconfig. The statistics IPC interface is always available
but will return 0 for all fields if the feature is disabled.

Change-Id: I5e51270902f0bcfb71f9b6d23861bc3994518e2c