Enable auto MACs per default
The switch shall set MACs automatically unless disabled. This way
clients get a unique MAC and do not need to set a default one which
would be the same for each client. Clients in L4Re do not have a good
way of selecting a unique MAC, thus the network should do it.
This behavior can be disabled via the new -M switch. -m is retained for
now, ignored and issues a warning.

Change-Id: Ic032b28ab93e2b298d5ef7c12860e3e33538fbf1