Use random MAC instead of hard coded values
If the "-m" option is passed, we set a MAC address for all switch
ports. Unless the user passes the "mac=" option when creating the port,
use a random address to be minimize the possibility of collisions.

Previously, we used to have hard coded addresses that only depend on
the (internal) port index. This is dangerous as those addresses are not
exactly globally unique.

Change-Id: I8954e98156d50d255ab19fa4e49702a385571586