Add missing initialization for Virtio_switch::_monitor member
The Virtio_switch::_monitor pointer was not initialized to nullptr,
which resulted in undefined behavior, as we use this uninitialized
member. This worked by pure luck until now, because apparently the
memory that got allocated for the Virtio_switch still was all zeroes,
since it was the initial allocation of that memory.

Change-Id: Id21afd1f196b007e7b7d60fb0acddd940fd3898c